backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
sentinelhub
supabase
sentinelsat
pydantic
